requirements.txt
Django==5.2.4
psycopg2-binary>=2.9.9
pymongo>=4.6.0
python-decouple>=3.8